Fix spigot - forge syncing
