+++ platforms/spigot/src/main/java/com/pg85/otg/spigot/networking/BiomeSettingSyncWrapper.java
@@ -29,23 +29,27 @@ public class BiomeSettingSyncWrapper
 	public BiomeSettingSyncWrapper(ByteBuf buffer)
 	{
 		this.fogDensity = buffer.readFloat();
+		byte size;
 
 		List<ColorThreshold> grassColors = new ArrayList<>();
-		for (int i = 0; i < buffer.readInt(); i++)
+		size = buffer.readByte();
+		for (int i = size; i > 0; i--)
 		{
 			grassColors.add(new ColorThreshold(buffer.readInt(), buffer.readFloat()));
 		}
 		this.grassColorControl = new SimpleColorSet(grassColors);
 
 		List<ColorThreshold> foliageColors = new ArrayList<>();
-		for (int i = 0; i < buffer.readInt(); i++)
+		size = buffer.readByte();
+		for (int i = size; i > 0; i--)
 		{
 			foliageColors.add(new ColorThreshold(buffer.readInt(), buffer.readFloat()));
 		}
 		this.foliageColorControl = new SimpleColorSet(foliageColors);
 
 		List<ColorThreshold> waterColors = new ArrayList<>();
-		for (int i = 0; i < buffer.readInt(); i++)
+		size = buffer.readByte();
+		for (int i = size; i > 0; i--)
 		{
 			waterColors.add(new ColorThreshold(buffer.readInt(), buffer.readFloat()));
 		}
@@ -77,21 +81,21 @@ public class BiomeSettingSyncWrapper
 	{
 		buffer.writeFloat(this.fogDensity);
 
-		buffer.writeByte((byte)this.grassColorControl.getLayers().size());
+		buffer.writeByte((byte) this.grassColorControl.getLayers().size());
 		for (ColorThreshold color : this.grassColorControl.getLayers())
 		{
 			buffer.writeInt(color.getColor());
 			buffer.writeFloat(color.getMaxNoise());
 		}
 
-		buffer.writeByte((byte)this.foliageColorControl.getLayers().size());
+		buffer.writeByte((byte) this.foliageColorControl.getLayers().size());
 		for (ColorThreshold color : this.foliageColorControl.getLayers())
 		{
 			buffer.writeInt(color.getColor());
 			buffer.writeFloat(color.getMaxNoise());
 		}
 
-		buffer.writeByte((byte)this.waterColorControl.getLayers().size());
+		buffer.writeByte((byte) this.waterColorControl.getLayers().size());
 		for (ColorThreshold color : this.waterColorControl.getLayers())
 		{
 			buffer.writeInt(color.getColor());
